Update oldLady.js
This program needs reworking. It runs out of control, I can't see why. The purpose of it is to alter lines from the poem, "There was an Old Lady who Swallowed a Fly" , pulling the words from an array that changed every stanza. Then these changes, along with one new line, would be tacked onto the lines from the previous stanza in order to form the new one.